Moved attributes-related code to separate module.
